chore: rename to quill keyboard
